refactor: remove bandaid for v3 specific error code
