+++ packages/nocodb-sdk/src/lib/globals.ts
@@ -151,14 +151,10 @@ export enum NcErrorType {
   API_TOKEN_NOT_ALLOWED = 'API_TOKEN_NOT_ALLOWED',
   WORKSPACE_NOT_FOUND = 'WORKSPACE_NOT_FOUND',
   BASE_NOT_FOUND = 'BASE_NOT_FOUND',
-  BASE_NOT_FOUNDV3 = 'BASE_NOT_FOUNDV3',
   SOURCE_NOT_FOUND = 'SOURCE_NOT_FOUND',
   TABLE_NOT_FOUND = 'TABLE_NOT_FOUND',
-  TABLE_NOT_FOUNDV3 = 'TABLE_NOT_FOUNDV3',
   VIEW_NOT_FOUND = 'VIEW_NOT_FOUND',
-  VIEW_NOT_FOUNDV3 = 'VIEW_NOT_FOUNDV3',
   FIELD_NOT_FOUND = 'FIELD_NOT_FOUND',
-  FIELD_NOT_FOUNDV3 = 'FIELD_NOT_FOUNDV3',
   RECORD_NOT_FOUND = 'RECORD_NOT_FOUND',
   GENERIC_NOT_FOUND = 'GENERIC_NOT_FOUND',
   HOOK_NOT_FOUND = 'HOOK_NOT_FOUND',
@@ -169,7 +165,6 @@ export enum NcErrorType {
   INVALID_PAGE_VALUE = 'INVALID_PAGE_VALUE',
   INVALID_LIMIT_VALUE = 'INVALID_LIMIT_VALUE',
   INVALID_FILTER = 'INVALID_FILTER',
-  INVALID_FILTERV3 = 'INVALID_FILTERV3',
   INVALID_SHARED_VIEW_PASSWORD = 'INVALID_SHARED_VIEW_PASSWORD',
   INVALID_ATTACHMENT_JSON = 'INVALID_ATTACHMENT_JSON',
   NOT_IMPLEMENTED = 'NOT_IMPLEMENTED',
@@ -196,13 +191,6 @@ export enum NcErrorType {
   BASE_USER_ERROR = 'BASE_USER_ERROR',
 }
 
-export const NcErrorTypeMap = {
-  [NcErrorType.TABLE_NOT_FOUNDV3]: 'TABLE_NOT_FOUND',
-  [NcErrorType.BASE_NOT_FOUNDV3]: 'BASE_NOT_FOUND',
-  [NcErrorType.VIEW_NOT_FOUNDV3]: 'VIEW_NOT_FOUND',
-  [NcErrorType.FIELD_NOT_FOUNDV3]: 'FIELD_NOT_FOUND',
-  [NcErrorType.INVALID_FILTERV3]: 'INVALID_FILTER',
-};
 export const LongTextAiMetaProp = 'ai';
 
 export const NO_SCOPE = 'nc';
@@ -227,9 +215,9 @@ export const DURATION_TYPE_MAP = {
   'h:mm:ss.s': 2,
   'h:mm:ss.ss': 3,
   'h:mm:ss.sss': 4,
-}
+};
 
-export const CURRENT_USER_TOKEN = '@me'
+export const CURRENT_USER_TOKEN = '@me';
 
 type Roles = OrgUserRoles | ProjectRoles | WorkspaceUserRoles;
