+++ packages/nocodb/src/filters/global-exception/global-exception.filter.ts
@@ -6,7 +6,6 @@ import hash from 'object-hash';
 import {
   NcApiVersion,
   NcErrorType,
-  NcErrorTypeMap,
   NcSDKError,
   NcSDKErrorV2,
   BadRequest as SdkBadRequest,
@@ -230,7 +229,7 @@ export class GlobalExceptionFilter implements ExceptionFilter {
       return response.status(422).json({ msg: exception.message });
     } else if (exception instanceof NcSDKErrorV2) {
       return response.status(exception.getStatus?.() ?? 422).json({
-        error: NcErrorTypeMap[exception.errorType] ?? exception.errorType,
+        error: exception.errorType,
         message: exception.message,
       });
     } else if (exception instanceof TestConnectionError) {
@@ -239,7 +238,7 @@ export class GlobalExceptionFilter implements ExceptionFilter {
         .json({ msg: exception.message, sql_code: exception.sql_code });
     } else if (exception instanceof NcBaseErrorv2) {
       return response.status(exception.code).json({
-        error: NcErrorTypeMap[exception.error] ?? exception.error,
+        error: exception.error,
         message: exception.message,
         details: exception.details,
       });
