+++ packages/nocodb/src/middlewares/extract-ids/extract-ids.middleware.ts
@@ -102,11 +102,7 @@ export class ExtractIdsMiddleware implements NestMiddleware, CanActivate {
       );
 
       if (!base) {
-        if (context.api_version === NcApiVersion.V3) {
-          NcError.baseNotFoundV3(params.baseId ?? params.baseName);
-        } else {
-          NcError.baseNotFound(params.baseId ?? params.baseName);
-        }
+        NcError.get(context).baseNotFound(params.baseId ?? params.baseName);
       }
 
       if (base) {
@@ -125,11 +121,7 @@ export class ExtractIdsMiddleware implements NestMiddleware, CanActivate {
           );
 
           if (!model) {
-            if (context.api_version === NcApiVersion.V3) {
-              NcError.tableNotFoundV3(params.tableId || params.modelId);
-            } else {
-              NcError.tableNotFound(req.params.tableName);
-            }
+            NcError.get(context).tableNotFound(req.params.tableName);
           }
 
           req.ncSourceId = model?.source_id;
@@ -146,11 +138,7 @@ export class ExtractIdsMiddleware implements NestMiddleware, CanActivate {
       });
 
       if (!model) {
-        if (context.api_version === NcApiVersion.V3) {
-          NcError.tableNotFoundV3(params.tableId || params.modelId);
-        } else {
-          NcError.tableNotFound(params.tableId || params.modelId);
-        }
+        NcError.get(context).tableNotFound(params.tableId || params.modelId);
       }
 
       req.ncBaseId = model.base_id;
@@ -337,11 +325,7 @@ export class ExtractIdsMiddleware implements NestMiddleware, CanActivate {
       });
 
       if (!model) {
-        if (context.api_version === NcApiVersion.V3) {
-          NcError.tableNotFoundV3(params.tableId || params.modelId);
-        } else {
-          NcError.tableNotFound(req.body.fk_model_id);
-        }
+        NcError.get(context).tableNotFound(req.body.fk_model_id);
       }
 
       req.ncBaseId = model.base_id;
@@ -365,11 +349,7 @@ export class ExtractIdsMiddleware implements NestMiddleware, CanActivate {
       });
 
       if (!model) {
-        if (context.api_version === NcApiVersion.V3) {
-          NcError.tableNotFoundV3(params.tableId || params.modelId);
-        } else {
-          NcError.tableNotFound(req.query?.fk_model_id);
-        }
+        NcError.get(context).tableNotFound(req.query?.fk_model_id);
       }
 
       req.ncBaseId = model.base_id;
