+++ packages/nocodb/src/services/data-table.service.ts
@@ -6,7 +6,7 @@ import {
   ViewTypes,
 } from 'nocodb-sdk';
 import { validatePayload } from 'src/helpers';
-import { NcApiVersion } from 'nocodb-sdk';
+import type { NcApiVersion } from 'nocodb-sdk';
 import type { LinkToAnotherRecordColumn } from '~/models';
 import type { NcContext } from '~/interface/config';
 import { nocoExecute } from '~/utils';
@@ -311,11 +311,7 @@ export class DataTableService {
   ) {
     const model = await Model.get(context, param.modelId);
     if (!model) {
-      if (context.api_version === NcApiVersion.V3) {
-        NcError.tableNotFoundV3(param.modelId);
-      } else {
-        NcError.tableNotFound(param.modelId);
-      }
+      NcError.get(context).tableNotFound(param.modelId);
     }
 
     if (param.baseId && model.base_id !== param.baseId) {
@@ -327,11 +323,7 @@ export class DataTableService {
     if (param.viewId) {
       view = await View.get(context, param.viewId);
       if (!view || (view.fk_model_id && view.fk_model_id !== param.modelId)) {
-        if (context.api_version === NcApiVersion.V3) {
-          NcError.viewNotFoundV3(param.viewId);
-        } else {
-          NcError.viewNotFound(param.viewId);
-        }
+        NcError.get(context).viewNotFound(param.viewId);
       }
     }
